Move deal list reorder endpoint from v2 to v1
