+++ backend/src/main/java/com/skapp/community/crmplanner/controller/v2/CrmDealControllerV2.java
@@ -4,7 +4,6 @@ import com.skapp.community.common.payload.response.ResponseEntityDto;
 import com.skapp.community.crmplanner.payload.request.CrmDealCreateRequestDto;
 import com.skapp.community.crmplanner.payload.request.CrmDealEditRequestDto;
 import com.skapp.community.crmplanner.payload.request.CrmDealFilterDto;
-import com.skapp.community.crmplanner.payload.request.CrmDealListReorderRequestDto;
 import com.skapp.community.crmplanner.service.v2.CrmDealServiceV2;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -59,15 +58,6 @@ public class CrmDealControllerV2 {
 	@PreAuthorize("hasAnyRole('ROLE_CRM_SALES_REPRESENTATIVE')")
 	public ResponseEntity<ResponseEntityDto> getDealById(@PathVariable Long id) {
 		ResponseEntityDto response = crmDealService.getDealById(id);
-		return new ResponseEntity<>(response, HttpStatus.OK);
-	}
-
-	@Operation(summary = "Reorder a deal in the list view",
-			description = "Repositions a deal within the table (list) view order between two neighbours.")
-	@PatchMapping("/reorder")
-	@PreAuthorize("hasAnyRole('ROLE_CRM_SALES_REPRESENTATIVE')")
-	public ResponseEntity<ResponseEntityDto> reorderDealInList(@RequestBody CrmDealListReorderRequestDto requestDto) {
-		ResponseEntityDto response = crmDealService.reorderDealInList(requestDto);
 		return new ResponseEntity<>(response, HttpStatus.OK);
 	}
