+++ backend/src/main/java/com/skapp/community/crmplanner/service/v2/CrmDealServiceV2.java
@@ -4,7 +4,6 @@ import com.skapp.community.common.payload.response.ResponseEntityDto;
 import com.skapp.community.crmplanner.payload.request.CrmDealCreateRequestDto;
 import com.skapp.community.crmplanner.payload.request.CrmDealEditRequestDto;
 import com.skapp.community.crmplanner.payload.request.CrmDealFilterDto;
-import com.skapp.community.crmplanner.payload.request.CrmDealListReorderRequestDto;
 
 public interface CrmDealServiceV2 {
 
@@ -17,7 +16,5 @@ public interface CrmDealServiceV2 {
 	ResponseEntityDto createDeal(CrmDealCreateRequestDto requestDto);
 
 	ResponseEntityDto editDeal(Long id, CrmDealEditRequestDto requestDto);
-
-	ResponseEntityDto reorderDealInList(CrmDealListReorderRequestDto requestDto);
 
 }
